+++ lib/components/AdapterStateTile/AdapterStateTile.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_blue/flutter_blue.dart';
+
+class AdapterStateTile extends StatelessWidget {
+  const AdapterStateTile({Key key, @required this.state}) : super(key: key);
+
+  final BluetoothState state;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: Colors.redAccent,
+      child: ListTile(
+        title: Text(
+          'Bluetooth adapter is ${state.toString().substring(15)}',
+          style: Theme.of(context).primaryTextTheme.subtitle1,
+        ),
+        trailing: Icon(
+          Icons.error,
+          color: Theme.of(context).primaryTextTheme.subtitle1.color,
+        ),
+      ),
+    );
+  }
+}
