+++ lib/components/DescriptorTile/DescriptorTile.dart
@@ -1,0 +1,54 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_blue/flutter_blue.dart';
+
+class DescriptorTile extends StatelessWidget {
+  final BluetoothDescriptor descriptor;
+  final VoidCallback onReadPressed;
+  final VoidCallback onWritePressed;
+
+  const DescriptorTile(
+      {Key key, this.descriptor, this.onReadPressed, this.onWritePressed})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: <Widget>[
+          Text('Descriptor'),
+          Text('0x${descriptor.uuid.toString().toUpperCase().substring(4, 8)}',
+              style: Theme.of(context)
+                  .textTheme
+                  .body1
+                  .copyWith(color: Theme.of(context).textTheme.caption.color))
+        ],
+      ),
+      subtitle: StreamBuilder<List<int>>(
+        stream: descriptor.value,
+        initialData: descriptor.lastValue,
+        builder: (c, snapshot) => Text(snapshot.data.toString()),
+      ),
+      trailing: Row(
+        mainAxisSize: MainAxisSize.min,
+        children: <Widget>[
+          IconButton(
+            icon: Icon(
+              Icons.file_download,
+              color: Theme.of(context).iconTheme.color.withOpacity(0.5),
+            ),
+            onPressed: onReadPressed,
+          ),
+          IconButton(
+            icon: Icon(
+              Icons.file_upload,
+              color: Theme.of(context).iconTheme.color.withOpacity(0.5),
+            ),
+            onPressed: onWritePressed,
+          )
+        ],
+      ),
+    );
+  }
+}
