+++ lib/components/ServiceTile/ServiceTile.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_blue/flutter_blue.dart';
+import 'package:glove_reader/components/CharacteristicTile/CharacteristicTile.dart';
+
+class ServiceTile extends StatelessWidget {
+  final BluetoothService service;
+  final List<CharacteristicTile> characteristicTiles;
+
+  const ServiceTile({Key key, this.service, this.characteristicTiles})
+      : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    if (characteristicTiles.length > 0) {
+      return ExpansionTile(
+        title: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: <Widget>[
+            Text('Service'),
+            Text('0x${service.uuid.toString().toUpperCase().substring(4, 8)}',
+                style: Theme.of(context)
+                    .textTheme
+                    .body1
+                    .copyWith(color: Theme.of(context).textTheme.caption.color))
+          ],
+        ),
+        children: characteristicTiles,
+      );
+    } else {
+      return ListTile(
+        title: Text('Service'),
+        subtitle:
+            Text('0x${service.uuid.toString().toUpperCase().substring(4, 8)}'),
+      );
+    }
+  }
+}
